refactor(context-budget): 统一模型族与 ContextBudgetProbe 快照构建
- 新增 ModelFamily 统一上下文上限与 tokenizer 编码选择
- 以 ContextBudgetProbe + build_context_budget_snapshot 替代旧 from_items API
- 探测项用 Cow 借用、OpenAI 投影支持 Borrow，stack_order 改为 usize
- 投影失败改为 ContextBudgetProjectionErrorKind 结构化 payload，UI 短文案
- session context_budget 模块承载 overlay payload，删除 terminal-ui payload.rs

+++ crates/runtime-domain/src/context_budget.rs
@@ -29,7 +29,7 @@ impl SegmentKind {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ContextSegment {
     pub kind: SegmentKind,
-    pub stack_order: u16,
+    pub stack_order: usize,
     pub estimated_tokens: usize,
 }
